🗃️ refactor: add migrations from v3 to v4
