+++ src/store/global/slices/settings/selectors/modelProvider.ts
@@ -1,11 +1,8 @@
-import { produce } from 'immer';
-
 import {
   AnthropicProvider,
   BedrockProvider,
   GoogleProvider,
   GroqProvider,
-  LOBE_DEFAULT_MODEL_LIST,
   MistralProvider,
   MoonshotProvider,
   OllamaProvider,
@@ -17,65 +14,11 @@ import {
   ZhiPuProvider,
   filterEnabledModels,
 } from '@/config/modelProviders';
-import { ChatModelCard, ModelProviderCard } from '@/types/llm';
-import { parseModelString } from '@/utils/parseModels';
+import { ModelProviderCard } from '@/types/llm';
+import { transformToChatModelCards } from '@/utils/parseModels';
 
 import { GlobalStore } from '../../../store';
 import { currentSettings } from './settings';
-
-/**
- * Extract a special method to process chatModels
- * @param modelConfig
- * @param defaultChartModels
- */
-const processChatModels = (
-  modelConfig: ReturnType<typeof parseModelString>,
-  defaultChartModels = OpenAIProvider.chatModels,
-): ChatModelCard[] => {
-  let chatModels = modelConfig.removeAll ? [] : defaultChartModels;
-
-  // 处理移除逻辑
-  if (!modelConfig.removeAll) {
-    chatModels = chatModels.filter((m) => !modelConfig.removed.includes(m.id));
-  }
-
-  return produce(chatModels, (draft) => {
-    // 处理添加或替换逻辑
-    for (const toAddModel of modelConfig.add) {
-      // first try to find the model in LOBE_DEFAULT_MODEL_LIST to confirm if it is a known model
-      const knownModel = LOBE_DEFAULT_MODEL_LIST.find((model) => model.id === toAddModel.id);
-
-      // if the model is known, update it based on the known model
-      if (knownModel) {
-        const modelInList = draft.find((model) => model.id === toAddModel.id);
-
-        // if the model is already in chatModels, update it
-        if (modelInList) {
-          // if (modelInList.hidden) delete modelInList.hidden;
-          modelInList.enabled = true;
-          if (toAddModel.displayName) modelInList.displayName = toAddModel.displayName;
-        } else {
-          // if the model is not in chatModels, add it
-          draft.push({
-            ...knownModel,
-            displayName: toAddModel.displayName || knownModel.displayName || knownModel.id,
-            enabled: true,
-          });
-        }
-      } else {
-        // if the model is not in LOBE_DEFAULT_MODEL_LIST, add it as a new custom model
-        draft.push({
-          ...toAddModel,
-          displayName: toAddModel.displayName || toAddModel.id,
-          enabled: true,
-          functionCall: true,
-          // isCustom: true,
-          vision: true,
-        });
-      }
-    }
-  });
-};
 
 // const azureModelList = (s: GlobalStore): ModelProviderCard => {
 //   const azure = azureConfig(s);
@@ -87,31 +30,22 @@ const processChatModels = (
 
 /**
  * define all the model list of providers
- * @param s
  */
 const providerModelList = (s: GlobalStore): ModelProviderCard[] => {
-  const openaiModelConfig = parseModelString(s.serverConfig.customModelName);
+  const openaiChatModels = transformToChatModelCards(s.serverConfig.customModelName);
 
-  const openaiChatModels = processChatModels(openaiModelConfig);
-
-  const ollamaModelConfig = parseModelString(s.serverConfig.languageModel?.ollama?.customModelName);
-
-  const ollamaChatModels = processChatModels(ollamaModelConfig, OllamaProvider.chatModels);
-
-  const openRouterModelConfig = parseModelString(
-    s.serverConfig.languageModel?.openrouter?.customModelName,
+  const ollamaChatModels = transformToChatModelCards(
+    s.serverConfig.languageModel?.ollama?.customModelName,
+    OllamaProvider.chatModels,
   );
 
-  const openrouterChatModels = processChatModels(
-    openRouterModelConfig,
+  const openrouterChatModels = transformToChatModelCards(
+    s.serverConfig.languageModel?.openrouter?.customModelName,
     OpenRouterProvider.chatModels,
   );
 
-  const togetheraiModelConfig = parseModelString(
+  const togetheraiChatModels = transformToChatModelCards(
     currentSettings(s).languageModel.togetherai.customModelName,
-  );
-  const togetheraiChatModels = processChatModels(
-    togetheraiModelConfig,
     TogetherAIProvider.chatModels,
   );
